+++ Backend/Models/GradeType.cs
@@ -9,7 +9,7 @@ public class GradeType
 {
     public int GradeTypeID { get; set; }
     public string Name { get; set; } // e.g., "Homework", "Quiz", "Midterm", etc.
-    public int? MaxGrade { get; set; } = 0;
+    public decimal? MaxGrade { get; set; } = 0;
     public bool IsActive { get; set; } = true;
 
     public ICollection<MonthlyGrade> MonthlyGrades { get; set; }
